Exception sur produit gratuit

+++ src/main/resources/schema.sql
@@ -2,6 +2,6 @@ CREATE TABLE product
 (
     id        INT PRIMARY KEY,
     nom       VARCHAR(255) NOT NULL,
-    prix      INT         NOT NULL,
-    prix_achat INT         NOT NULL
+    prix      DOUBLE         NOT NULL,
+    prix_achat DOUBLE         NOT NULL
 );
